requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
psycopg2
Jinja2==2.9.4
MarkupSafe==0.23
Werkzeug==0.11.15
click==6.7
itsdangerous==0.24
wsgiref==0.1.2
requests
beautifulsoup4
six==1.10.0
PyJWT==1.4.2
pyparsing==2.2.0
packaging==16.8
python-bcrypt==0.3.1
# ----------------------
# Production Server
# ----------------------
gunicorn==19.6.0
